Move `assert_vectors_equal` to `util/testing`
